Rename youtube_downloader module
