+++ docs/releases/changelog.md
@@ -7,6 +7,7 @@ Le format est inspiré de [Keep a Changelog](https://keepachangelog.com/fr/1.1.0
 ## [Unreleased]
 - Ajout de la variable d'environnement `PYDL_LOG_LEVEL` pour contrôler la verbosité des logs.
 - Nouvelle dataclass `ProgressOptions` et gestionnaires `ProgressBarHandler`/`VerboseProgressHandler` pour suivre l'avancement des téléchargements.
+- Renommage de `youtube_downloader.py` en `legacy_utils.py`.
 
 ## [0.1.0] - 2025-06-06
 ### Added
