+++ program_youtube_downloader/__init__.py
@@ -1,6 +1,7 @@
 """Public exports for program_youtube_downloader."""
 from .exceptions import PydlError, DownloadError, ValidationError
-from .utils import clear_screen, program_break_time
+# re-export legacy utilities
+from .legacy_utils import clear_screen, program_break_time
 
 __all__ = [
     "PydlError",
